Move feed selection into its own file without changing the behavior.
